move to single binary
